+++ src/main/java/com/joyent/manta/cosbench/MantaStorage.java
@@ -5,7 +5,6 @@ import com.intel.cosbench.api.storage.StorageException;
 import com.intel.cosbench.config.Config;
 import com.intel.cosbench.log.Logger;
 import com.joyent.manta.client.MantaClient;
-import com.joyent.manta.client.MantaHttpHeaders;
 import com.joyent.manta.client.MantaMetadata;
 import com.joyent.manta.config.ChainedConfigContext;
 import com.joyent.manta.config.DefaultsConfigContext;
@@ -15,11 +14,13 @@ import com.joyent.manta.config.SystemSettingsConfigContext;
 import com.joyent.manta.cosbench.config.CosbenchMantaConfigContext;
 import com.joyent.manta.exception.MantaClientHttpResponseException;
 import com.joyent.manta.exception.MantaErrorCode;
+import com.joyent.manta.http.MantaHttpHeaders;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
 /**
  * Manta implementation of the COSBench {@link com.intel.cosbench.api.storage.StorageAPI}.
@@ -30,7 +31,7 @@ public class MantaStorage extends NoneStorage {
     /**
      * Hardcoded directory in Manta in which all benchmark files are stored.
      */
-    private static final String COSBENCH_BASE_DIR = "stor/cosbench";
+    private static final String DEFAULT_COSBENCH_BASE_DIR = "stor/cosbench";
 
     /**
      * The default number of maximum HTTP connections at one time to the
@@ -49,14 +50,31 @@ public class MantaStorage extends NoneStorage {
     private String currentTestDirectory;
 
     /**
-     * Flag indicating whether or not to send files using chunked encoding.
-     */
-    private Boolean chunked;
-
-    /**
      * Number of copies of object to store.
      */
     private Integer durabilityLevel;
+
+    /**
+     * Flag that toggles chunked transfer encoding.
+     */
+    private boolean chunked;
+
+    /**
+     * Flag indicating that logging is enabled.
+     */
+    private boolean logging;
+
+    /**
+     * Number of sections in which to download files. If greater than one,
+     * then multiple HTTP Range requests will be used to assemble the file.
+     */
+    private int sections;
+
+    /**
+     * Size of the object being benchmarked - used only with HTTP range request
+     * benchmarks.
+     */
+    private Integer objectSize;
 
     @Override
     public void init(final Config config, final Logger logger) {
@@ -77,17 +95,46 @@ public class MantaStorage extends NoneStorage {
                 new SystemSettingsConfigContext(),
                 cosbenchConfig);
 
-        this.chunked = cosbenchConfig.useChunking();
         this.durabilityLevel = cosbenchConfig.getDurabilityLevel();
+        this.logging = cosbenchConfig.logging();
+        this.sections = cosbenchConfig.getNumberOfSections();
+        this.objectSize = cosbenchConfig.getObjectSize();
 
-        logger.info(String.format("Client configuration: %s", context));
+        if (cosbenchConfig.chunked() == null) {
+            if (logging) {
+                logger.info("Chunked mode is disabled");
+            }
+
+            this.chunked = false;
+        } else {
+            final String status;
+            if (cosbenchConfig.chunked()) {
+                status = "enabled";
+            } else {
+                status = "disabled";
+            }
+
+            if (logging) {
+                logger.info("Chunked mode is " + status);
+            }
+            this.chunked = cosbenchConfig.chunked();
+        }
+
+        if (logging) {
+            logger.info(String.format("Client configuration: %s",
+                    context));
+        }
 
         try {
             client = new MantaClient(context);
 
+            final String baseDir = Objects.toString(
+                    cosbenchConfig.getBaseDirectory(),
+                    DEFAULT_COSBENCH_BASE_DIR);
+
             // We rely on COSBench properly cleaning up after itself.
             currentTestDirectory = String.format("%s/%s",
-                    context.getMantaHomeDirectory(), COSBENCH_BASE_DIR);
+                    context.getMantaHomeDirectory(), baseDir);
 
             client.putDirectory(currentTestDirectory, true);
 
@@ -101,33 +148,45 @@ public class MantaStorage extends NoneStorage {
             throw new StorageException(e);
         }
 
-        logger.debug("Manta client has been initialized");
+        if (logging) {
+            logger.debug("Manta client has been initialized");
+        }
     }
 
     @Override
     public void createContainer(final String container, final Config config) {
-        super.createContainer(container, config);
+        if (logging) {
+            logger.info("Performing PUT at /{}", container);
+        }
 
         try {
             client.putDirectory(directoryOfContainer(container));
         } catch (Exception e) {
-            logger.error("Error creating container", e);
+            if (logging) {
+                logger.error("Error creating container", e);
+            }
         }
     }
 
     @Override
     public void deleteContainer(final String container, final Config config) {
-        super.deleteContainer(container, config);
+        if (logging) {
+            logger.info("Performing DELETE at /{}", container);
+        }
 
         try {
             client.deleteRecursive(directoryOfContainer(container));
         } catch (MantaClientHttpResponseException e) {
             if (!e.getServerCode().equals(MantaErrorCode.RESOURCE_NOT_FOUND_ERROR)) {
-                logger.error("Error error deleting object", e);
+                if (logging) {
+                    logger.error("Error error deleting object", e);
+                }
                 throw new StorageException(e);
             }
         } catch (Exception e) {
-            logger.error("Error deleting container", e);
+            if (logging) {
+                logger.error("Error deleting container", e);
+            }
             throw new StorageException(e);
         }
     }
@@ -138,22 +197,27 @@ public class MantaStorage extends NoneStorage {
                              final InputStream data,
                              final long length,
                              final Config config) {
-        super.createObject(container, object, data, length, config);
+        if (logging) {
+            logger.info("Performing PUT at /{}/{}", container, object);
+        }
 
         final String path = pathOfObject(container, object);
+        final long contentLength;
+
+        if (chunked) {
+            contentLength = -1L;
+        } else {
+            contentLength = length;
+        }
+
+        MantaHttpHeaders headers = new MantaHttpHeaders();
 
         try {
-            MantaHttpHeaders headers = new MantaHttpHeaders();
-
-            if (chunked != null && chunked) {
-                headers.setContentEncoding("chunked");
-            }
-
             if (durabilityLevel != null) {
                 headers.setDurabilityLevel(durabilityLevel);
             }
 
-            client.put(path, data, headers);
+            client.put(path, data, contentLength, headers, null);
         } catch (MantaClientHttpResponseException e) {
             // This is a fall-back in the weird cases where COSBench doesn't
             // do things in the right order.
@@ -161,7 +225,7 @@ public class MantaStorage extends NoneStorage {
                 try {
                     String dir = directoryOfContainer(container);
                     client.putDirectory(dir, true);
-                    client.put(path, data);
+                    client.put(path, data, contentLength, headers, null);
                 } catch (IOException ioe) {
                     throw new StorageException(ioe);
                 }
@@ -169,7 +233,10 @@ public class MantaStorage extends NoneStorage {
                 throw new StorageException(e);
             }
         } catch (Exception e) {
-            logger.error("Error error creating object", e);
+            if (logging) {
+                logger.error("Error error creating object", e);
+            }
+
             throw new StorageException(e);
         }
     }
@@ -177,33 +244,64 @@ public class MantaStorage extends NoneStorage {
     @Override
     public void deleteObject(final String container, final String object,
                              final Config config) {
-        super.deleteObject(container, object, config);
+        if (logging) {
+            logger.info("Performing DELETE at /{}/{}", container, object);
+        }
 
         try {
             String path = pathOfObject(container, object);
             client.delete(path);
         } catch (MantaClientHttpResponseException e) {
             if (!e.getServerCode().equals(MantaErrorCode.RESOURCE_NOT_FOUND_ERROR)) {
-                logger.error("Error error deleting object", e);
+                if (logging) {
+                    logger.error("Error error deleting object", e);
+                }
                 throw new StorageException(e);
             }
         } catch (Exception e) {
-            logger.error("Error error deleting object", e);
+            if (logging) {
+                logger.error("Error error deleting object", e);
+            }
             throw new StorageException(e);
         }
     }
 
     @Override
     public InputStream getObject(final String container, final String object, final Config config) {
-        super.getObject(container, object, config);
+        final InputStream objectStream;
 
         try {
             final String path = pathOfObject(container, object);
-            return client.getAsInputStream(path);
+
+            if (sections == 1) {
+                if (logging) {
+                    logger.info("Performing GET at /{}/{}",
+                            container, object);
+                }
+                objectStream = client.getAsInputStream(path);
+            } else if (objectSize != null) {
+                if (logging) {
+                    logger.info("Performing GET with HTTP byte range at /{}/{}", container, object);
+                }
+
+                int size = this.objectSize;
+                objectStream = new RangeJoiningInputStream(path, client, size, sections);
+            } else {
+                String msg = "[object-size] must be set when [no-of-http-range-sections] is set";
+
+                if (logging) {
+                    logger.error(msg);
+                }
+                throw new StorageException(msg);
+            }
         } catch (Exception e) {
-            logger.error("Error error getting object", e);
+            if (logging) {
+                logger.error("Error error getting object", e);
+            }
             throw new StorageException(e);
         }
+
+        return objectStream;
     }
 
     @Override
@@ -211,7 +309,9 @@ public class MantaStorage extends NoneStorage {
                                   final String object,
                                   final Map<String, String> map,
                                   final Config config) {
-        super.createMetadata(container, object, map, config);
+        if (logging) {
+            logger.info("Performing POST at /{}/{}", container, object);
+        }
 
         try {
             String path = pathOfObject(container, object);
@@ -225,7 +325,9 @@ public class MantaStorage extends NoneStorage {
             MantaMetadata metadata = new MantaMetadata(prefixedMap);
             client.putMetadata(path, metadata);
         } catch (Exception e) {
-            logger.error("Error error creating metadata", e);
+            if (logging) {
+                logger.error("Error error creating metadata", e);
+            }
             throw new StorageException(e);
         }
     }
@@ -234,27 +336,31 @@ public class MantaStorage extends NoneStorage {
     protected Map<String, String> getMetadata(final String container,
                                               final String object,
                                               final Config config) {
-        super.getMetadata(container, object, config);
+        if (logging) {
+            logger.info("Performing HEAD at /{}/{}", container, object);
+        }
 
         try {
             String path = pathOfObject(container, object);
             return client.head(path).getMetadata();
         } catch (Exception e) {
-            logger.error("Error error getting metadata", e);
+            if (logging) {
+                logger.error("Error error getting metadata", e);
+            }
             throw new StorageException(e);
         }
     }
 
     @Override
     public void dispose() {
-        super.dispose();
-
         try {
             if (client != null) {
                 client.close();
             }
         } catch (Exception e) {
-            logger.warn("Error when attempting to close Manta client", e);
+            if (logging) {
+                logger.warn("Error when attempting to close Manta client", e);
+            }
         }
 
         client = null;
@@ -262,7 +368,7 @@ public class MantaStorage extends NoneStorage {
 
     @Override
     public void abort() {
-        super.abort();
+        client.closeQuietly();
     }
 
     /**