+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2016-04-06
+### Changed
+ - Upgraded Java Manta driver library to 3.0.0.
+ - Removed chunked encoding configuration option.
+ - Added HTTP range testing support.
+ 
 ## [1.0.7] - 2016-11-17
 ### Changed
  - Upgraded Java Manta driver library.